Move build under the bundle command
* porter bundle build
* porter build
